All services and models were rewrite on Kotlin

+++ src/main/java/tk/hiddenname/smarthome/exception/ProcessorExistsException.kt
@@ -4,4 +4,4 @@ import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
 
 @ResponseStatus(value = HttpStatus.CONFLICT)
-class ProcessorExistsException(id: Int) : Exception("Processor with id '$id' has been already existed")
+class ProcessorExistsException(id: Long) : Exception("Processor with id '$id' has been already existed")
