+++ src/main/java/tk/hiddenname/smarthome/exception/ProcessorNotFoundException.kt
@@ -4,4 +4,4 @@ import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
 
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
-class ProcessorNotFoundException(id: Int) : Exception("The processor with id '$id' not found")
+class ProcessorNotFoundException(id: Long) : Exception("The processor with id '$id' not found")
