+++ src/main/java/tk/hiddenname/smarthome/exception/TriggerExistsException.kt
@@ -4,4 +4,4 @@ import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
 
 @ResponseStatus(value = HttpStatus.CONFLICT)
-class TriggerExistsException(id: Int) : Exception("The trigger with id '$id' have been already existed")
+class TriggerExistsException(id: Long) : Exception("The trigger with id '$id' have been already existed")
